Some updates to the TagController

+++ src/main/java/nl/cookplanner/controllers/TagController.java
@@ -5,18 +5,28 @@ import java.util.Optional;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import lombok.extern.slf4j.Slf4j;
 import nl.cookplanner.domain.Tag;
+import nl.cookplanner.execptions.ErrorResponse;
+import nl.cookplanner.execptions.TagAlreadyExistsException;
+import nl.cookplanner.execptions.TagNotFoundException;
 import nl.cookplanner.services.TagService;
 
+@Slf4j
 @RestController
+@RequestMapping("tag")
+@CrossOrigin(origins = "http://localhost:4200")
 public class TagController {
 
 	private final TagService tagService;
@@ -25,39 +35,61 @@ public class TagController {
 		this.tagService = tagService;
 	}
 
-	@GetMapping("/tags")
-	public List<Tag> getTags() {
-		return tagService.listAll();
+	@GetMapping("/list")
+	public ResponseEntity<List<Tag>> getTags() {
+		return ResponseEntity.ok(tagService.listAll());
 	}
-	
-	@GetMapping("/tags/{id}")
+
+	@GetMapping("/{id}")
 	public ResponseEntity<Tag> get(@PathVariable Long id) {
 		Optional<Tag> optionalTag = tagService.get(id);
 		if (optionalTag.isPresent()) {
-			return new ResponseEntity<>(optionalTag.get(), HttpStatus.OK);
+			return ResponseEntity.ok(optionalTag.get());
 		} else {
-			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+			throw new TagNotFoundException("Categorie kan niet worden gevonden");
 		}
 	}
 	
-	@PostMapping("/tags")
-	public void add(@RequestBody Tag tag) {
-		tagService.save(tag);
+	@PostMapping("/create")
+	public ResponseEntity<Tag> create(@RequestBody Tag tag) throws TagAlreadyExistsException {
+		return ResponseEntity.ok(tagService.save(tag));
+		
 	}
 	
-	@PutMapping("/tags/{id}")
+	@PutMapping("/update/{id}")
 	public ResponseEntity<Tag> update(@RequestBody Tag tag, @PathVariable Long id) {
 		Optional<Tag> optionalTag = tagService.get(id);
 		if (optionalTag.isPresent()) {
-			tagService.save(tag);
-			return new ResponseEntity<>(HttpStatus.OK);
+			return ResponseEntity.ok(tagService.save(tag));
 		} else {
-			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+			throw new TagNotFoundException("Categorie kan niet worden gevonden");
 		}
 	}
 	
-	@DeleteMapping("tags/{id}")
-	public void delete(@PathVariable Long id) {
-		tagService.delete(id);
+	@DeleteMapping("/delete/{id}")
+	public ResponseEntity<Tag> delete(@PathVariable Long id) {
+		// TODO: Build additional check if category is in use and handle accordingly
+		Optional<Tag> optionalTag = tagService.get(id);
+		if (optionalTag.isPresent()) {
+			tagService.delete(id);
+			return ResponseEntity.ok(optionalTag.get());
+		} else {
+			throw new TagNotFoundException("Categorie kan niet worden gevonden");
+		}
+		
 	}
+	
+	@ExceptionHandler({TagNotFoundException.class})
+    public ResponseEntity<ErrorResponse> tagNotFound(TagNotFoundException ex){
+
+        return new ResponseEntity<>(
+            new ErrorResponse(ex.getMessage(), 404) , HttpStatus.NOT_FOUND);
+    }
+	
+	@ExceptionHandler({TagAlreadyExistsException.class})
+    public ResponseEntity<ErrorResponse> tagNotFound(TagAlreadyExistsException ex){
+
+        return new ResponseEntity<>(
+            new ErrorResponse(ex.getMessage(), 400) , HttpStatus.BAD_REQUEST);
+    }
 }
